% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{df2jira}
\alias{df2jira}
\title{Table converting function}
\usage{
df2jira(df)
}
\arguments{
\item{df}{This is the dataframe you want to convert to jira format}
}
\description{
This funcition converts your dataframe to jira format. It prints the jira code to your console.
}
\examples{
df2jira(df)
}
\keyword{dataframe,}
\keyword{jira,}
\keyword{table}

