% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{world_tour}
\alias{world_tour}
\title{Attempt to apply query to multiple TSO DWHs}
\usage{
world_tour assumes you have ODBC setups such that each ODBC is named like TSO_<country> eg. TSO_DE.
}
\arguments{
\item{query}{SQL query in a text format}

\item{countries}{Countries you like to query from. Text format. Default is c("DE","PL","FR","BR","CZ","NCSA_ES","US","NL","UK","RU","EMEA_ES","IT","GR","RO","CN","MENA")}
}
\description{
This function allows you to fetch data from all the TSO databases using your query.
}
\examples{
dd <- world_tour("SELECT date(datedim) date, dr FROM agg_revenue")
it <- world_tour("SELECT date(datedim) date, dr FROM agg_revenue","IT")
}
\keyword{database,}
\keyword{odbc}
\keyword{query,}
\keyword{sql,}

