% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{write.excel}
\alias{write.excel}
\title{write.excel}
\usage{
write.excel(data, sheet = "data", file = "data.xlsx", row.names = F)
}
\arguments{
\item{data}{data you like to save in Excel format}

\item{sheet}{Sheet name (default "data")}

\item{file}{File name (default "data.xlsx")}

\item{row.names}{If you like to include row names, logical (default F)}
}
\description{
Write data in xlsx format just as easily as write_csv
}
\examples{
data(iris)
write.excel(iris)
}
\keyword{excel,}
\keyword{xlsx}

